requirements.txt
# ML Models
torch>=2.0.0
transformers>=4.30.0
sentencepiece>=0.1.99

# Messaging - Core
python-telegram-bot>=20.0

# Messaging - Optional Platforms
discord.py>=2.3.0              # Discord support
slack-bolt>=1.18.0             # Slack support
webwhatsapi>=2.0.0             # WhatsApp support (requires selenium)
selenium>=4.15.0               # Required for WhatsApp

# Utilities
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0
pydantic>=2.0.0
pyyaml>=6.0

# Dev
pytest>=7.4.0
black>=23.0.0
